requirements.txt
typer>=0.12.5
rich>=13.7.1
PyYAML>=6.0.2